moved analyze tests
